Consolidate db code into database package
