+++ depobs/scanner/models/nodejs.py
@@ -5,7 +5,7 @@ from dataclasses import asdict, dataclass, field
 import enum
 from typing import Dict, Tuple, Sequence, List, Optional, Generator, Union
 
-from depobs.scanner.serialize_util import (
+from depobs.util.serialize_util import (
     extract_fields,
     get_in,
     JSONPath,
