+++ depobs/scanner/models/pipeline.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass, field
 from typing import AbstractSet, Callable, Optional
 
 from depobs.scanner.graph_util import NODE_ID_FORMATS, NODE_LABEL_FORMATS, GROUP_ATTRS
-from depobs.scanner.serialize_util import identity_serializer
+from depobs.util.serialize_util import identity_serializer
 
 
 def add_infile_and_outfile(
