+++ depobs/scanner/models/rust.py
@@ -3,7 +3,7 @@ from dataclasses import asdict, dataclass, field
 import enum
 from typing import Dict, Tuple, Sequence, List, Optional
 
-from depobs.scanner.serialize_util import extract_fields, get_in
+from depobs.util.serialize_util import extract_fields, get_in
 
 log = logging.getLogger(__name__)
