+++ depobs/scanner/pipelines/postprocess.py
@@ -26,7 +26,7 @@ from typing import (
 import typing
 
 from depobs.scanner.graph_util import npm_packages_to_networkx_digraph, get_graph_stats
-from depobs.scanner.serialize_util import (
+from depobs.util.serialize_util import (
     get_in,
     extract_fields,
     extract_nested_fields,
