+++ depobs/scanner/pipelines/rust_changelog.py
@@ -18,7 +18,7 @@ from typing import (
     AsyncGenerator,
 )
 
-from depobs.scanner.serialize_util import (
+from depobs.util.serialize_util import (
     get_in,
     extract_fields,
 )
