+++ setup.cfg
@@ -23,11 +23,10 @@ directory = /tmp/htmlcov
 files =
       depobs/scanner/*.py,
       depobs/scanner/clients/*.py,
-      depobs/scanner/db/*.py,
       depobs/scanner/docker/*.py,
       depobs/scanner/models/*.py,
       depobs/scanner/pipelines/__init__.py,
-      depobs/website/models.py,
+      depobs/database/models.py,
       depobs/worker/scoring.py
 
 python_version = 3.8
